+++ src/cli/tc.rs
@@ -59,7 +59,10 @@ pub async fn run() -> Result<()> {
                         .into_iter()
                         .collect::<Result<Vec<_>>>()?;
 
-                    validate_operations(&blueprint, ops).await.to_cli_result()?;
+                    validate_operations(&blueprint, ops)
+                        .await
+                        .to_cli_result()
+                        .map_err(|e| e.message("Invalid Operation".into()))?;
 
                     blueprint.validate_rest_apis().await.to_cli_result()?;
